+++ src/javascript/0045-ActionList.js
@@ -22,16 +22,22 @@ s.ActionList.prototype._verify = function(obj) {
     return (obj instanceof s.Action);
 }
 
-s.ActionList.prototype.addAction = function(text, carryOutFunc, checkFunc) {
+s.ActionList.prototype.addAction = function(
+    text, carryOutFunc, checkFunc, index
+) {
     /*
     This method is for adding a simple action with one outcome to the
     list. It creates an outcome that executes `carryOutFunc`, adds it as
     the single outcome to a new action with link text `text` and check
     function `checkFunc`. The `checkFunc` parameter defaults to true.
+
+    If an `index` parameter is given, will insert the action at the
+    given index, otherwise will push the action.
     */
+    index = index || this.length();
     var outcome = new s.Outcome(carryOutFunc);
     var action = new s.Action(text, checkFunc);
     action.push(outcome);
-    this.push(action);
+    this.insert(index, action);
     return this;
 }